[Parley] refactor: Split ScriptBrowserController.cs (#707)
Split 638-line controller into focused components:
- ScriptBrowserController.cs (299 lines): Script browser window management
- ParameterBrowserController.cs (290 lines): Parameter browser and declaration caching
- ScriptPreviewService.cs (88 lines): Script preview loading and display

Tests: All 500 tests pass ✅
